+++ apps/wiki/tasks.py
@@ -1,10 +1,12 @@
 import logging
 
 from django.conf import settings
+from django.contrib.auth.models import User
 from django.contrib.sites.models import Site
 from django.core.cache import cache
 from django.core.exceptions import ValidationError
 from django.core.mail import send_mail, mail_admins
+from django.db import transaction
 from django.dispatch import receiver
 from django.template import Context, loader
 
@@ -131,3 +133,53 @@ def build_json_data_handler(sender, instance, **kwargs):
     except:
         logging.error('JSON metadata build task failed',
                       exc_info=True)
+
+
+@task
+@transaction.commit_on_success
+def move_page(locale, slug, new_slug, email):
+    try:
+        user = User.objects.get(email=email)
+    except User.DoesNotExist:
+        logging.error('Page move failed: no user with email address %s' % email)
+        return
+    try:
+        doc = Document.objects.get(locale=locale, slug=slug)
+    except Document.DoesNotExist:
+        message = """
+        Page move failed.
+
+        Move was requested for document with slug %(slug)s in locale %(locale)s,
+        but no such document exists.
+        """ % {'slug': slug, 'locale': locale}
+        logging.error(message)
+        send_mail('Page move failed', message, settings.DEFAULT_FROM_ADDRESS,
+                  [user.email])
+                  
+        return
+    try:
+        doc._move_tree(new_slug, user=user)
+    except Exception as e:
+        message = """
+        Page move failed.
+
+        Move was requested for document with slug %(slug)s in locale %(locale)s,
+        but could not be completed. The following error was raised:
+
+        %(info)
+        """ % {'slug': slug, 'locale': locale, 'info': e}
+        logging.error(message)
+        send_mail('Page move failed', message, settings.DEFAULT_FROM_ADDRESS,
+                  [user.email])
+
+
+    message = """
+    Page move completed.
+
+    The move requested for the document with slug %(slug)s in locale
+    %(locale)s, and all its children, has been completed.
+
+    You can now view this document at its new slug, %(new_slug)s.
+    """ % {'slug': slug, 'locale': locale, 'new_slug': new_slug}
+    send_mail('Page move completed', message, settings.DEFAULT_FROM_ADDRESS,
+              [user.email])
